// src/app/models/order.ts
export class Order {
    // orderId: number;
    // empId: number;
    // empName: string;
    // email: string;
    // itemRequested: string;
    // qty: number;
    // shippingAddress: string;
    // shippedDate: string;
    // phnNo: number;
    // orderDate: string;
    // isRejectedByAdmin: number;

    address: string;
    deliveryDate: Date;
    email: string;
    empFirstName: string;
    empLastName: string;
    isRejectedByAdmin: number;
    orderDate: Date;
    orderId: number;
    phoneNumber: number;
    productCategory: string;
    productImageUrl: string;
    productName: string;
    productPrice: number;
    productQty: number;
    productRating: number;
    qty: number;
    shippedDate: Date;
}